feat: bash execution of ensemble run
In order to run the ensemble calculations in parallel or to distribute them among several machines the bash script run01.sh is a first step into this direction. Instead of a big internal ensemble loop this script runs the cross section determination file by file, setting the material ID via the variable "drsdlD".